+++ src/nemmstrategy_shortterm/BuyStrategy1Tactic.java
@@ -35,22 +35,22 @@ public class BuyStrategy1Tactic extends GenericTactic {
 		discount = d;	
 	}
 	
-	private BuyOffer creatBuyOfferone(double expectedprice, int physicalposition) {
+	private BuyOffer creatBuyOfferone(double expectedprice, double physicalposition) {
 		BuyOffer ret = new BuyOffer();
-		ret.numberofcert = (int) (shareboughtatdiscount*(-physicalposition)); //-As the phisical position of buyer would in most cases be negative, but as the offer only has positive numbers. 
+		ret.numberofcert = (shareboughtatdiscount*(-physicalposition)); //-As the phisical position of buyer would in most cases be negative, but as the offer only has positive numbers. 
 		ret.price = (1-discount)*expectedprice;
 		return ret;
 		}
 	
 
-	private BuyOffer creatBuyOffertwo(double expectedprice, int physicalposition) {
+	private BuyOffer creatBuyOffertwo(double expectedprice, double physicalposition) {
 		BuyOffer ret = new BuyOffer();
-		ret.numberofcert = (-physicalposition) -( (int) (shareboughtatdiscount*(-physicalposition))); //rest of the monthly production bought at expected price.
+		ret.numberofcert = (-physicalposition) -( (shareboughtatdiscount*(-physicalposition))); //rest of the monthly production bought at expected price.
 		ret.price = expectedprice;
 		return ret;
 		}
 	
-	public void updatetacticbuyoffers(double expectedprice, int physicalposition) {
+	public void updatetacticbuyoffers(double expectedprice, double physicalposition) {
 		if (physicalposition > 0){
 			physicalposition = 0;} //To ensure that we dont get crazy bids.  
 		tacticbuyoffers.clear();
