+++ src/nemmstrategy_shortterm/PAUtilityMethod.java
@@ -21,7 +21,7 @@ public class PAUtilityMethod extends GenericUtilityMethod{
 	}
 	
 	//Takes in the given values and calculates the producers utility just based on what maximizes value of certificates sold. 
-	public Double calculateutility(double marketprice, ArrayList<BuyOffer> b, ArrayList<SellOffer> s, int physicalposition, double shareofmarginaltoffersold, double shareofmarginalofferbought) {
+	public Double calculateutility(double marketprice, ArrayList<BuyOffer> b, ArrayList<SellOffer> s, double physicalposition, double shareofmarginaltoffersold, double shareofmarginalofferbought) {
 		double ret;
 		ret = UpdatePhysicalPosition.returnsoldvolume(s, marketprice, shareofmarginaltoffersold).getSoldInSTMcert() * marketprice;
 		return ret;
